Add files via upload
One of my first university web projects, where I've used HTML/CSS, JavaScript and Bootstrap. It's not optimized and made on my old laptop's resolution. But it has plenty of pages and I've learned a lot while working on it. Also many of icons and images are made by myself in photoshop.